requirements.txt
pyyaml>=6.0
jsonschema>=4.20
pytest>=7.4
ruff>=0.1.0
mypy>=1.7
